Mon, Mar 14, 2022, 9:40 PM +00:00

+++ filter-more/helpers.c
@@ -9,7 +9,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     BYTE red;
     BYTE green;
     BYTE blue;
-    BYTE avg;
+    int avg;
 
     for (int i = 0; i < height; i++)
     {
